Move AdminNavBar into common admin folder
