+++ app/frontend/src/components/common/admin/AdminNavBar.vue
