+++ app/frontend/src/views/ForestrySectorOpScreening.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script>
-import AdminNavBar from '@/components/common/AdminNavBar.vue';
+import AdminNavBar from '@/components/common/admin/AdminNavBar.vue';
 import forestrySectorOpScreeningForm from '@/store/modules/forestrysectoropscreening/forestrySectorOpScreeningForm.js';
 import { AppClients, FormNames } from '@/utils/constants';
